KAS-2992 cleanup: configuration.rb > government-domains [pub feedback]
