+++ lib/linked_db.rb
@@ -14,6 +14,8 @@ module LinkedDB
 
   BASE_BACKOFF_SECONDS = 1
   def self.query(query)
+    log.debug(query)
+    
     max_tries = 5
     (1..max_tries).each do |try|
       begin
